requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
ecdsa==0.19.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-restx==1.3.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.0
python-jose==3.3.0
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
rsa==4.9
six==1.16.0
Werkzeug==3.1.3
